Fix type 32 variables in DE and FR
Some of the variables that refer to other objects have 0xFF byte in "postfix" of such variable. Because of that the conversion script was trying to show this as a variable option, which in turn broke the string.
This behavior is fixed in the script. It ignores now the next 0xFF after the first one in type 32 and 33 variables.